move schema type to client
